Fixed variable name and working directory
It is $BUDGET_ACCOUNTS instead of $BUDGET_ACCOUNT (even though it always
contains at most one budget account).

It is now more clear that the other two variables are to be resolving
path names.

Now the working directory is set correctly in every code block (because
it is apparantly needed in every code block).