Do not assume existence of dta/{simres,restart}
dta/simres and dta/restart do not contain any tracked files and thus
existence cannot be assumed.  One cannot expect realpath(1) to work on
non-existing path, even as it does work on the GNU implementation.  The
fix is to append the basename(1) to the realpath(1) of the dirname(1).
Only caveat would be users using a different basename, but that is
really fringe.